+++ bcbio/pipeline/datadict.py
@@ -58,6 +58,7 @@ LOOKUPS = {
     "oncofuse_file": {"keys": ["oncofuse_file"]},
     "split_bam": {"keys": ["split_bam"]},
     "vrn_file": {"keys": ["vrn_file"]},
+    "variant_regions": {"keys": ["config", "algorithm", "variant_regions"]},
     "deduped_bam": {"keys": ["deduped_bam"]},
     "align_bam": {"keys": ["align_bam"]}
 }